// src/types/UserTypes.ts
export type UserType = {
  userId: string;
  theme: string;
  font: string;
  language: string;
  fname: string;
  lname: string;
  phone: string;
  email: string;
  username: string;
  password: string;
};

export type SettingType = {
  theme: string;
  font: string;
  language: string;
};

export type LoginType = {
  username: string;
  password: string;
};